# docs/swagger.yaml
definitions:
...
  dto.BannerCardDTO:
    properties:
      image_url:
        type: string
      target_id:
        type: string
      target_type:
        type: string
      target_url:
        type: string
    type: object
...
  dto.CartLineDTO:
    properties:
      average_price:
        $ref: '#/definitions/dto.MoneyDTO'
      item_id:
        type: string
      pack_label:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      variant_id:
        type: string
    type: object
  dto.CartResponse:
    properties:
      cart_id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.CartLineDTO'
        type: array
      platforms:
        items:
          $ref: '#/definitions/dto.PlatformBreakdownDTO'
        type: array
    type: object
...
  dto.CategoryCardDTO:
    properties:
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.CityDTO:
    properties:
      city_id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  dto.CityItem:
    description: A serviceable city. The lat/lng + default_pincode are the location
    properties:
      city_id:
        description: Unique identifier for the city
        type: string
      default_pincode:
        description: Pincode used as the default QuickCommerce location anchor
        type: string
      enabled:
        description: Whether the city is serviceable
        type: boolean
      lat:
        description: Latitude of the city's QuickCommerce location anchor
        type: number
      lng:
        description: Longitude of the city's QuickCommerce location anchor
        type: number
      name:
        description: Display name of the city
        type: string
      slug:
        description: URL-safe name
        type: string
      state:
        description: State the city sits in
        type: string
    type: object
...
  dto.HomeResponse:
    properties:
      banners:
        items:
          $ref: '#/definitions/dto.BannerCardDTO'
        type: array
      categories:
        items:
          $ref: '#/definitions/dto.CategoryCardDTO'
        type: array
      top_stores:
        items:
          $ref: '#/definitions/dto.PlatformDTO'
        type: array
      trending_items:
        items:
          $ref: '#/definitions/dto.ProductCardDTO'
        type: array
    type: object
...
  dto.MoneyDTO:
    properties:
      display:
        type: string
      paise:
        type: integer
    type: object
...
  dto.PlatformBreakdownDTO:
    properties:
      available_item_ids:
        items:
          type: string
        type: array
      available_total:
        $ref: '#/definitions/dto.MoneyDTO'
      delivery_eta_minutes:
        type: integer
      is_cheapest:
        type: boolean
      item_total:
        $ref: '#/definitions/dto.MoneyDTO'
      logo_url:
        type: string
      platform_code:
        type: string
      platform_name:
        type: string
      unavailable_item_ids:
        items:
          type: string
        type: array
    type: object
  dto.PlatformDTO:
    properties:
      code:
        type: string
      delivery_eta_text:
        type: string
      logo_url:
        type: string
      name:
        type: string
    type: object
...
  dto.PlatformPriceChipDTO:
    properties:
      available:
        type: boolean
      deep_link:
        type: string
      logo_url:
        type: string
      mrp:
        $ref: '#/definitions/dto.MoneyDTO'
      platform_code:
        type: string
      platform_name:
        type: string
      price:
        $ref: '#/definitions/dto.MoneyDTO'
    type: object
...
  dto.ProductCardDTO:
    properties:
      brand_name:
        type: string
      default_variant_id:
        type: string
      image_url:
        type: string
      name:
        type: string
      product_id:
        type: string
      starting_price:
        $ref: '#/definitions/dto.MoneyDTO'
    type: object
  dto.ProductCardListDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ProductCardDTO'
        type: array
      meta:
        $ref: '#/definitions/query.Meta'
    type: object
...
  dto.ProductDetailDTO:
    properties:
      brand_name:
        type: string
      category_id:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      product_id:
        type: string
      similar:
        items:
          $ref: '#/definitions/dto.ProductCardDTO'
        type: array
      variants:
        items:
          $ref: '#/definitions/dto.VariantDetailDTO'
        type: array
    type: object
...
  dto.Response:
    description: Generic response envelope wrapping every API response.
    properties:
      code:
        description: Machine-readable error/status code, present on failures
        type: string
      data:
        description: Response payload, shape depends on the endpoint
      message:
        description: Human-readable message describing the result
        type: string
      status:
        description: Outcome of the request, e.g. "success" or "failed"
        type: string
    type: object
...
  dto.VariantDetailDTO:
    properties:
      average_price:
        $ref: '#/definitions/dto.MoneyDTO'
      pack_label:
        description: '"500 gm"'
        type: string
      platform_prices:
        items:
          $ref: '#/definitions/dto.PlatformPriceChipDTO'
        type: array
      unit_price:
        description: '"₹72/100gm"'
        type: string
      variant_id:
        type: string
    type: object
  query.Meta:
    properties:
      page:
        type: integer
      pages:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  v1.CreateCityRequest:
    properties:
      default_pincode:
        type: string
      enabled:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      state:
        type: string
    required:
    - default_pincode
    - lat
    - lng
    - name
    type: object
...
  v1.DeleteCityRequest:
    properties:
      city_id:
        type: string
    required:
    - city_id
    type: object
...
  v1.UpdateCityRequest:
    properties:
      city_id:
        type: string
      default_pincode:
        type: string
      enabled:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      state:
        type: string
    required:
    - city_id
    type: object
...
  v1.addCartItemRequest:
    properties:
      quantity:
        type: integer
      variant_id:
        type: string
    required:
    - variant_id
    type: object
...
  v1.quantityRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
...
  v1.wishlistRequest:
    properties:
      variant_id:
        type: string
    required:
    - variant_id
    type: object
info:
  contact:
    name: API Support
  description: This is the API documentation for the Grocerics backend service.
  title: Grocerics API
  version: "1.1"
paths:
...
  /v1/cart:
    get:
      description: The cart plus, for every enabled platform, the available/not-available
        split and totals.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get cart with per-platform breakdown
      tags:
      - consumer
  /v1/cart/items:
    post:
      consumes:
      - application/json
      description: Add a variant (upserts quantity if already present). Returns the
        updated cart.
      parameters:
      - description: variant + quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.addCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Add an item to the cart
      tags:
      - consumer
  /v1/cart/items/{id}:
    delete:
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Remove a cart item
      tags:
      - consumer
    patch:
      consumes:
      - application/json
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      - description: new quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.quantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update a cart item's quantity
      tags:
      - consumer
...
  /v1/categories/{id}/products:
    get:
      description: Paginated grid of product cards for a category in the user's city.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: page size, max 100 (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductCardListDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Products in a category (PLP)
      tags:
      - consumer
...
  /v1/cities:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete City Request
        in: body
        name: DeleteCityRequest
        required: true
        schema:
          $ref: '#/definitions/v1.DeleteCityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a city
      tags:
      - cities
    get:
      description: Cities the app serves — used by the location picker.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CityDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: List serviceable cities
      tags:
      - consumer
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update City Request
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateCityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CityItem'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a city
      tags:
      - cities
    post:
      consumes:
      - application/json
      parameters:
      - description: Create City Request
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/v1.CreateCityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CityItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a serviceable city
      tags:
      - cities
...
  /v1/deals:
    get:
      description: Products that have a discounted platform price (mrp > price) in
        the user's city.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProductCardDTO'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Top Deals
      tags:
      - consumer
  /v1/home:
    get:
      description: Banners, top stores (platforms), trending categories and trending
        items for the user's city.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HomeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Home screen
      tags:
      - consumer
...
  /v1/products/{id}:
    get:
      description: Variants with per-platform prices, average price, unit price, and
        similar products.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductDetailDTO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Product detail (PDP)
      tags:
      - consumer
  /v1/search:
    get:
      description: Full-text search over the Groceric catalog (min 2 chars). Never
        hits the platform APIs.
      parameters:
      - description: search term (min 2 chars)
        in: query
        name: q
        required: true
        type: string
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: page size, max 100 (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductCardListDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Search products (curated catalog)
      tags:
      - consumer
...
  /v1/wishlist:
    get:
      description: Same per-platform available/not-available shape as the cart.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get wishlist with per-platform breakdown
      tags:
      - consumer
    post:
      consumes:
      - application/json
      parameters:
      - description: variant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.wishlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Add a variant to the wishlist
      tags:
      - consumer
  /v1/wishlist/{variantId}:
    delete:
      parameters:
      - description: Variant ID
        in: path
        name: variantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Remove a variant from the wishlist
      tags:
      - consumer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"